Move controller to Management area
