+++ EComWeb/Areas/Identity/Pages/Account/Manage/ExternalLogins.cshtml.cs
@@ -122,7 +122,7 @@ namespace EComWeb.Areas.Identity.Pages.Account.Manage
             var info = await _signInManager.GetExternalLoginInfoAsync(userId);
             if (info == null)
             {
-                throw new InvalidOperationException($"Unexpected error occurred loading external login info.");
+                throw new InvalidOperationException("Unexpected error occurred loading external login info.");
             }
 
             var result = await _userManager.AddLoginAsync(user, info);
