+++ EComWeb/Areas/Management/Controllers/SpecificationsController.cs
@@ -9,9 +9,10 @@ using ECom.DataAccess.Data;
 using ECom.Models;
 using Microsoft.AspNetCore.Authorization;
 
-namespace EComWeb.Controllers
+namespace EComWeb.Areas.Management.Controllers
 {
     [Authorize(Roles = "Admin")]
+    [Area("Management")]
     public class SpecificationsController : Controller
     {
         private readonly ApplicationDbContext _context;
@@ -100,7 +101,7 @@ namespace EComWeb.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Processor,Storage,RAM,Screen,Camera,ProductId")] Specification specification)
+        public async Task<IActionResult> Edit(int id, [Bind("Id,Processor,Storage,RAM,Screen,Camera,ProductId")] Specification specification)
         {
             if (id != specification.Id)
             {
