+++ EComWeb/Areas/Management/Controllers/UserController.cs
@@ -1,0 +1,27 @@
+﻿using ECom.Models;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+
+namespace EComWeb.Areas.Management.Controllers
+{
+    [Authorize(Roles = "Admin")]
+    [Area("Management")]
+    public class UserController : Controller
+    {
+        private readonly UserManager<ApplicationUser> _userManager;
+
+        public UserController(UserManager<ApplicationUser> userManager)
+        {
+            _userManager = userManager;
+        }
+        // GET
+        public async Task<IActionResult> Index()
+        {
+            var currentUser = await _userManager.GetUserAsync(HttpContext.User);
+            var allUserExceptCurrentUser = await _userManager.Users.Where(a => a.Id != currentUser.Id).ToListAsync();
+            return View(allUserExceptCurrentUser);
+        }
+    }
+}
