+++ EComWeb/Areas/Management/Pages/Admin.cshtml
@@ -5,9 +5,9 @@
 }
 <h1>@ViewData["Title"]</h1>
 
-<a asp-controller="User">Users</a>
-<a asp-controller="Manufactures">Manufactures</a>
-<a asp-controller="Categories">Categories</a>
-<a asp-controller="Specifications">Specifications</a>
-<a asp-controller="Products">Products</a>
-<a asp-controller="Specifications">Specification</a>
+<a asp-area="Management" asp-controller="User">Users</a>
+<a asp-area="Management" asp-controller="Manufactures">Manufactures</a>
+<a asp-area="Management" asp-controller="Categories">Categories</a>
+<a asp-area="Management" asp-controller="Specifications">Specifications</a>
+<a asp-area="Management" asp-controller="Products">Products</a>
+<a asp-area="Management" asp-controller="Specifications">Specification</a>
