+++ EComWeb/Areas/Management/Views/Manufactures/Delete.cshtml
