+++ EComWeb/Areas/Management/Views/_ViewImports.cshtml
@@ -1,0 +1,5 @@
+@using Microsoft.AspNetCore.Identity
+@using ECom.Models
+EComWeb.Pages
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
+@namespace EComWeb.Areas.Management.Views
