+++ EComWeb/Views/_ViewImports.cshtml
@@ -1,0 +1,4 @@
+@using Microsoft.AspNetCore.Identity
+
+EComWeb.Pages
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
